Refactor: 패키지 구조 변경
- Chapter 2 stack의 패키지 구조 변경
