+++ README.md
@@ -443,7 +443,9 @@ if (books.containsKey(key)) {
 
 ### 백준 2812 크게 만들기
 * https://www.acmicpc.net/problem/2812
-* 자리값이 큰 숫자를 삭제하여 그 자리에 더 큰 숫자가 올 수 있다면 삭제하는 것이 이득이다. 앞에서부터 K개를 삭제하여 내림차순이 되도록 부분 수열을 만들고, 더 삭제할 수 없으면 그대로 사용한다는 아이디어를 떠올리지 못했다. 
+* 자리값이 큰 숫자를 삭제하여 그 자리에 더 큰 숫자가 올 수 있다면 삭제하는 것이 이득이다. 앞에서부터 K개를 삭제하여 내림차순이 되도록 부분 수열을 만들고, 더 삭제할 수 없으면 그대로 사용한다는 아이디어를 떠올리지 못했다.
+
+## Chapter 3. 재귀 #1 - 순열과 조합
 
 ## Part 5. 단기완성 알고리즘
 ## 완전 탐색(Brute Force)
